Add interactive status and description to tickets
A status field has been added to the tickets to quickly mark whether a ticket is completed or not. This status can be toggled by clicking on the status emoji. Additionally, the full description of a ticket can now be viewed by clicking anywhere on the respective ticket row. This information is fetched lazily to reduce initial load times. The information is displayed right there inside the row, and can be hidden again by reclicking on the row.

These changes were made to improve user experience by providing better visual feedback and reducing the need for navigating to a different page to see full information. This should encourage better and faster interactions. Minor UI improvements were also made.